+++ components/editor/validation/validation-overview.tsx
@@ -12,15 +12,17 @@ import { validationSettings } from "@/lib/state/validation-settings"
 import { useLayoutState } from "@/lib/state/layout-state"
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip"
 
+export interface IValidationOverviewProps {
+    entityId?: string
+    validationRunning?: boolean
+    size?: ComponentProps<typeof Button>["size"]
+}
+
 export const ValidationOverview = memo(function ValidationOverview({
     entityId,
     validationRunning,
     size
-}: {
-    entityId?: string
-    validationRunning?: boolean
-    size?: ComponentProps<typeof Button>["size"]
-}) {
+}: IValidationOverviewProps) {
     const [initiallyHidden, setInitiallyHidden] = useState(true)
 
     const validationStore = useValidationStore()
@@ -124,7 +126,7 @@ export const ValidationOverview = memo(function ValidationOverview({
 
     return (
         <div
-            className={`${validationRunning && validationResults.length > 0 ? "opacity-50" : ""} ${initiallyHidden ? "opacity-0" : ""} transition-opacity`}
+            className={`${validationRunning && validationResults.length > 0 ? "opacity-50" : ""} ${initiallyHidden ? "invisible" : ""} transition-opacity`}
         >
             <Popover>
                 <Tooltip delayDuration={500}>